add run_app_context function
this is an async generator function wrapped in a new class,
_AsyncContextManager. _AsyncContextManager is a placeholder until 3.7
when we will have contextlib.asynccontextmanager.

The new function allows the user to start an app and run other async
operations on the same loop while it is running like so:

    async with run_app_context(app, ..., loop=loop) as (app, uris):
        # do other async stuff, when we leave this block, the app will
        # be stopped
        await something

run_app_context has the same parameters as the old run_app function, but
loop must be given so we make sure that the app gets the same loop as
was used to run the async function in the example code above.

I have also rewriten run_app to call run_app_context so that only one is
maintained.